admin/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
psycopg2-binary==2.9.1
django-cors-headers==3.5.0
pika==1.1.0